CASMPET-3926: Initial implementation of the cray-kiali Helm chart
Adds the functionality to deploy Kiali via the kiali-operator
chart.

Version 1.25.0 is chosen because that's the latest version that's
been released that works with the version of Istio that we're
currently using. Istio is 1.7.8.

See https://kiali.io/documentation/latest/installation-guide/#_version_compatibility

The resource requests and limits are taken from what we're using for
Kiali now.

The DestinationRule and VirtualService are moved here from cray-istio
and there will be a PR to cray-istio so that cray-istio stops creating
those.
cray-istio-deploy will be changed to disable starting up Kiali.