CASMPET-5120: Switch to sysmgmt Prometheus, disable Jaeger
Kiali was no longer working since the removal of Prometheus from
the Istio deploy (CASMPET-4467). The Prometheus reference is
updated to the one in sysmgmt-health.

Also, I disabled tracing. This was disabled in Istio recently with
CASMPET-5096.

This commit also updates the version number. There's no other
changes in this repository and it's new for CSM 1.2.