CASMINST-3848: Add kiali image to annotation
The kiali:v1.28.1 image is used by the deployment created by the
kiali operator that's deployed by this chart, but it's not being
included in the build. This is because the deployment is created
by the operator and not by a deployment created in a template in
the chart so the tool that extracts images used by charts doesn't
find it.

The fix is to add the kiali image to the `artifacthub.io/images`
annotation. The tool that extracts images used by charts uses this
annotation.